- Added static_cast<int>() to get rid of a warning on   i686_Linux-2.2.18_g++-2.95_LEDA ("comparision between signed and   unsigned).

+++ Packages/Min_sphere_of_spheres_d/include/CGAL/Min_sphere_of_spheres_d_gcc2-95-2_fix.h
@@ -1287,7 +1287,7 @@ namespace std
 
 inline int std::stringbuf::sync()
 {
-  if((mode & ios::out) == 0)
+  if(static_cast<int>(mode & ios::out) == 0)
     return EOF;
 
   streamsize n = pptr() - pbase();
